requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.4.0
django-heroku==0.3.1
django-nvd3==0.9.7
djangorestframework==3.9.1
gunicorn==19.9.0
Jinja2==2.10
MarkupSafe==1.1.1
psycopg2==2.7.7
PyMySQL==0.9.3
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2018.9
six==1.12.0
Unidecode==1.0.23
whitenoise==4.1.2
